Rework regression testing.
This includes a number of changes:
 - Test directories (diffkemp, diffsysctl, diffkabi) are replaced by three
   testing scripts. These follow the functionality that used to be in the
   directories - testing single functions, sysctl parameters, and kernel module
   parameters.
 - All tests have common YAML spec files - there is a separate file for each
   pair of kernel versions.
 - Spec files allow to specify single functions, sysctls, and kernel module
   parameters. Each testing script only takes its relevant entries.
 - No source files are part of the repo anymore.

+++ tests/regression/task_spec.py
@@ -4,29 +4,78 @@ import os
 import shutil
 
 from diffkemp.config import Config
-from diffkemp.llvm_ir.build_llvm import LlvmKernelBuilder
+from diffkemp.llvm_ir.kernel_source import KernelSource
+from diffkemp.semdiff.result import Result
+
+
+specs_path = os.path.abspath("tests/regression/test_specs")
+tasks_path = os.path.abspath("tests/regression/kernel_modules")
+
+
+class FunctionSpec:
+    """"
+    Specification of a function in kernel along the modules/source files where
+    it is located and the expected result of the function comparison between
+    two kernel versions.
+    """
+    def __init__(self, name, result, old_module=None, new_module=None):
+        self.name = name
+        self.old_module = old_module
+        self.new_module = new_module
+        if result == "generic":
+            self.result = Result.Kind.NONE
+        else:
+            self.result = Result.Kind[result.upper()]
 
 
 class TaskSpec:
-    """Base class for task specification representing one testing scenario."""
-    def __init__(self, spec, tasks_path, task_name, module_dir):
-        self.old_kernel = spec["old_kernel"]
-        self.new_kernel = spec["new_kernel"]
+    """
+    Task specification representing testing scenario.
+    Contains a list of functions to be compared with DiffKemp during the test.
+    """
+    def __init__(self, spec, task_name, tasks_path, kernel_path):
+        self.old_kernel_dir = os.path.join(kernel_path, spec["old_kernel"])
+        self.new_kernel_dir = os.path.join(kernel_path, spec["new_kernel"])
         self.name = task_name
         self.task_dir = os.path.join(tasks_path, task_name)
-        self.debug = spec["debug"] if "debug" in spec else False
         if "control_flow_only" in spec:
             self.control_flow_only = spec["control_flow_only"]
         else:
             self.control_flow_only = False
 
-        # Create LLVM builders and configuration
-        self.old_builder = LlvmKernelBuilder(self.old_kernel, module_dir,
-                                             debug=self.debug, rebuild=True)
-        self.new_builder = LlvmKernelBuilder(self.new_kernel, module_dir,
-                                             debug=self.debug, rebuild=True)
-        self.config = Config(self.old_builder, self.new_builder, 120, False,
-                             self.control_flow_only, False, False, None)
+        # Create LLVM sources and configuration
+        self.old_kernel = KernelSource(self.old_kernel_dir, True)
+        self.new_kernel = KernelSource(self.new_kernel_dir, True)
+        self.config = Config(self.old_kernel, self.new_kernel, False,
+                             self.control_flow_only, False, None)
+
+        self.functions = dict()
+
+    def finalize(self):
+        """
+        Task finalization - should be called before the task is destroyed.
+        """
+        self.old_kernel.finalize()
+        self.new_kernel.finalize()
+
+    def add_function_spec(self, fun, result):
+        """Add a function comparison specification."""
+        self.functions[fun] = FunctionSpec(fun, result)
+
+    def build_modules_for_function(self, fun):
+        """
+        Build LLVM modules containing definition of the compared function in
+        both kernels.
+        """
+        # Since PyTest may share KernelSource objects among tasks, we need
+        # to explicitly initialize kernels.
+        self.old_kernel.initialize()
+        self.new_kernel.initialize()
+        mod_old = self.old_kernel.get_module_for_symbol(fun)
+        mod_new = self.new_kernel.get_module_for_symbol(fun)
+        self.functions[fun].old_module = mod_old
+        self.functions[fun].new_module = mod_new
+        return mod_old, mod_new
 
     def _file_name(self, suffix, ext, name=None):
         """
@@ -66,9 +115,68 @@ class TaskSpec:
 
         if not os.path.isfile(self.old_llvm_file(name)):
             shutil.copyfile(old_module.llvm, self.old_llvm_file(name))
-        if not os.path.isfile(self.old_src_file(name)):
+        if old_src and not os.path.isfile(self.old_src_file(name)):
             shutil.copyfile(old_src, self.old_src_file(name))
         if not os.path.isfile(self.new_llvm_file(name)):
             shutil.copyfile(new_module.llvm, self.new_llvm_file(name))
-        if not os.path.isfile(self.new_src_file(name)):
+        if new_src and not os.path.isfile(self.new_src_file(name)):
             shutil.copyfile(new_src, self.new_src_file(name))
+
+
+class SysctlTaskSpec(TaskSpec):
+    """
+    Task specification for test of sysctl comparison.
+    Extends TaskSpec by data variable and proc handler function.
+    """
+    def __init__(self, spec, task_name, tasks_path, kernel_path, data_var):
+        TaskSpec.__init__(self, spec, task_name, tasks_path, kernel_path)
+        self.data_var = data_var
+        self.proc_handler = None
+        self.old_sysctl_module = None
+        self.new_sysctl_module = None
+
+    def add_proc_handler(self, proc_handler, result):
+        """Add proc handler function to the spec."""
+        self.proc_handler = proc_handler
+        self.add_function_spec(proc_handler, result)
+
+    def get_proc_handler_spec(self):
+        """Retrieve specification of the proc handler function."""
+        return self.functions[self.proc_handler]
+
+    def add_data_var_function(self, fun, result):
+        """Add a new function using the data variable."""
+        self.add_function_spec(fun, result)
+
+    def build_sysctl_module(self):
+        """Build the compared sysctl modules into LLVM."""
+        self.old_sysctl_module = self.old_kernel.get_sysctl_module(self.name)
+        self.old_sysctl_module.parse_sysctls(self.name)
+        self.new_sysctl_module = self.new_kernel.get_sysctl_module(self.name)
+        self.new_sysctl_module.parse_sysctls(self.name)
+
+
+class ModuleParamSpec(TaskSpec):
+    """
+    Task specification for test of kernel module parameter comparison.
+    Extends TaskSpec by module and parameter specification.
+    """
+    def __init__(self, spec, dir, mod, param, tasks_path, kernel_path):
+        TaskSpec.__init__(self, spec, "{}-{}".format(mod, param), tasks_path,
+                          kernel_path)
+        self.dir = dir
+        self.mod = mod
+        self.param = param
+        self.old_module = None
+        self.new_module = None
+
+    def build_module(self):
+        """Build the compared kernel modules into LLVM."""
+        self.old_module = self.old_kernel.get_module_for_kernel_mod(self.dir,
+                                                                    self.mod)
+        self.new_module = self.new_kernel.get_module_for_kernel_mod(self.dir,
+                                                                    self.mod)
+
+    def get_param(self):
+        """Get the name of the global variable representing the parameter."""
+        return self.old_module.find_param_var(self.param)
